requirements.txt
aiohttp
aiofiles
asyncio
ffmpeg-python
pillow
git+https://github.com/pyrogram/pyrogram@master
py-tgcalls==0.8.1rc1
pytube
wget
python-dotenv
search-engine-parser
tgcrypto
youtube-search-python==1.5.1
youtube-search
speedtest-cli
cryptography
dnspython
future
googletrans
hurry.filesize
motor
psutil
pykeyboard
pyromod
requests
speedtest-cli
uvloop
git+https://github.com/itsunknown-12/yt-dl@master
telegraph
heroku3
hachoir
lyricsgenius
solidAPI==0.1.8.9
pafy==0.5.5
httpx>=0.14.2
